* fix postun scripts

+++ base/conf/yaffas-certificates/redhat/yaffas-certificates.spec
@@ -58,40 +58,42 @@ for PRODUCT in $PRODUCTS; do
 done
 
 %postun
-set -e
-CERTDIR="/opt/yaffas/etc/ssl/certs/"
-CERTS="exim webmin cyrus usermin ldap zarafa-webaccess zarafa-server zarafa-gateway zarafa-ical mppserver"
-for i in $CERTS; do
-	rm -f $CERTDIR/$i{,.key,.crt}
-done
-[ `find $CERTDIR -type f` ] || rm -Rf $CERTDIR # FIXME
-PRODUCTS="webmin usermin"
-for PRODUCT in $PRODUCTS; do
-	CONF="/opt/yaffas/etc/$PRODUCT/miniserv.conf"
-	TMPCONF="/tmp/miniserv.conf"
-	if [ -r $CONF ]; then
-		echo "Editing $PRODUCT"
-		KC="key cert"
-		for BBKC in $KC; do
-			if [ "$BBKC" = "cert" ]; then
-				FILEEXTENSION="crt";
-			else
-				FILEEXTENSION="key";
-			fi
-			if grep -q "^${BBKC}file" $CONF; then
-				sed -e "s/^${BBKC}file=.*/${BBKC}file=\/opt\/yaffas\/etc\/ssl\/certs\/$PRODUCT.${FILEEXTENSION}/" $CONF > $TMPCONF
-				# FIXME ^
-				rm -f $CONF
-				mv -f $TMPCONF $CONF
-				chmod 600 $CONF
-			else
-				echo "${BBKC}file=/opt/yaffas/etc/ssl/certs/$PRODUCT.${FILEEXTENSION}" >> $CONF
-				# FIXME ^ (alte config wiederherstellen)
-			fi
-		done
-	fi
-done
-rmdir --ignore-fail-on-non-empty /opt/yaffas/etc/ssl
+if [ $1 -eq 0 ]; then
+	set -e
+	CERTDIR="/opt/yaffas/etc/ssl/certs/"
+	CERTS="exim webmin cyrus usermin ldap zarafa-webaccess zarafa-server zarafa-gateway zarafa-ical mppserver"
+	for i in $CERTS; do
+		rm -f $CERTDIR/$i{,.key,.crt}
+	done
+	[ `find $CERTDIR -type f` ] || rm -Rf $CERTDIR # FIXME
+	PRODUCTS="webmin usermin"
+	for PRODUCT in $PRODUCTS; do
+		CONF="/opt/yaffas/etc/$PRODUCT/miniserv.conf"
+		TMPCONF="/tmp/miniserv.conf"
+		if [ -r $CONF ]; then
+			echo "Editing $PRODUCT"
+			KC="key cert"
+			for BBKC in $KC; do
+				if [ "$BBKC" = "cert" ]; then
+					FILEEXTENSION="crt";
+				else
+					FILEEXTENSION="key";
+				fi
+				if grep -q "^${BBKC}file" $CONF; then
+					sed -e "s/^${BBKC}file=.*/${BBKC}file=\/opt\/yaffas\/etc\/ssl\/certs\/$PRODUCT.${FILEEXTENSION}/" $CONF > $TMPCONF
+					# FIXME ^
+					rm -f $CONF
+					mv -f $TMPCONF $CONF
+					chmod 600 $CONF
+				else
+					echo "${BBKC}file=/opt/yaffas/etc/ssl/certs/$PRODUCT.${FILEEXTENSION}" >> $CONF
+					# FIXME ^ (alte config wiederherstellen)
+				fi
+			done
+		fi
+	done
+	rmdir --ignore-fail-on-non-empty /opt/yaffas/etc/ssl
+fi
 
 %files
 %defattr(-,root,root)
